Closes #927
