--- src/constants.ts
@@ -1,6 +1,0 @@
-export const projectInitDependencyErrorMessage = () =>
-  "Please run `spk project init` command before running this command to initialize the project.";
-export const projectInitCvgDependencyErrorMessage = () =>
-  "Please run `spk project init` and `spk project cvg` commands before running this command to initialize the project.";
-export const projectCvgDependencyErrorMessage = () =>
-  "Please run `spk project cvg` command before running this command to create a variable group.";
