+++ src/lib/constants.ts
@@ -1,2 +1,11 @@
 export const BUILD_SCRIPT_URL =
   "https://raw.githubusercontent.com/Microsoft/bedrock/master/gitops/azure-devops/build.sh";
+
+export const projectInitDependencyErrorMessage =
+  "Please run `spk project init` command before running this command to initialize the project.";
+
+export const projectInitCvgDependencyErrorMessage =
+  "Please run `spk project init` and `spk project cvg` commands before running this command to initialize the project.";
+
+export const projectCvgDependencyErrorMessage =
+  "Please run `spk project cvg` command before running this command to create a variable group.";
